Update subtraction scrolllist to optimize entry renders
