+++ client/src/js/subtraction/components/Item.js
@@ -1,35 +1,34 @@
 import React from "react";
+import { connect } from "react-redux";
 import { ClipLoader } from "halogenium";
 import { LinkContainer } from "react-router-bootstrap";
 import { Col, Row } from "react-bootstrap";
+import { get } from "lodash-es";
 
 import { Flex, FlexItem, Icon, ListGroupItem } from "../../base";
 
-export default function SubtractionItem ({ id, description, ready }) {
+const SubtractionItem = ({ entry }) => {
 
     let icon;
 
-    if (ready) {
+    if (entry.ready) {
         icon = <Icon name="check" bsStyle="success" />;
     } else {
         icon = <ClipLoader size="14px" color="#3c8786" />;
     }
 
     return (
-        <LinkContainer key={id} className="spaced" to={`/subtraction/${id}`}>
+        <LinkContainer key={entry.id} className="spaced" to={`/subtraction/${entry.id}`}>
             <ListGroupItem>
                 <Row>
-                    <Col xs={8} md={4}>
-                        <strong>{id}</strong>
+                    <Col xs={8} md={8}>
+                        <strong>{entry.id}</strong>
                     </Col>
-                    <Col xsHidden smHidden md={3} className="text-muted">
-                        {description}
-                    </Col>
-                    <Col xs={4} md={5}>
+                    <Col xs={4} md={4}>
                         <Flex alignItems="center" className="pull-right">
                             {icon}
                             <FlexItem pad>
-                                <strong>{ready ? "Ready" : "Importing"}</strong>
+                                <strong>{entry.ready ? "Ready" : "Importing"}</strong>
                             </FlexItem>
                         </Flex>
                     </Col>
@@ -37,4 +36,10 @@ export default function SubtractionItem ({ id, description, ready }) {
             </ListGroupItem>
         </LinkContainer>
     );
-}
+};
+
+const mapStateToProps = (state, props) => ({
+    entry: get(state, `subtraction.documents[${props.index}]`, null)
+});
+
+export default connect(mapStateToProps)(SubtractionItem);
